Module Authenticator.Param

type t

val ca_file : ?⁠allowed_hashes:Mirage_crypto.Hash.hash list -> ?⁠crls:Core.Filename.t -> Core.Filename.t -> unit -> t

val ca_dir : ?⁠allowed_hashes:Mirage_crypto.Hash.hash list -> ?⁠crls:Core.Filename.t -> Core.Filename.t -> unit -> t

val cert_fingerprints : Mirage_crypto.Hash.hash -> ([ `host ] Domain_name.t * string) list -> t

The fingerprint can be collected from a browser or by invoking an openssl command like 'openssl x509 -in <pem_file> -noout -fingerprint -sha256'

val time : unit -> Ptime.t option

Async programs often don't use Ptime_clock, so this is provided as a convenience function. Relies on Unix.gettimeofday.

val to_authenticator : time:(unit -> Ptime.t option) -> t -> Authenticator.t Async.Deferred.Or_error.t